GUI/requirements.txt
pyqt6
pyusb